# docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.JwtRequest:
    properties:
      jwt:
        type: string
    required:
    - jwt
    type: object
  models.JwtResponse:
    properties:
      jwt:
        type: string
    type: object
  models.UserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.UserResponse:
    properties:
      email:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Service for auth
  title: GiicooAuth
  version: "1.0"
paths:
  /check-token/:
    post:
      consumes:
      - application/json
      description: Check JWT token
      parameters:
      - description: Write JWT token
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.JwtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Check JWT Token
      tags:
      - auth
  /generate-token/:
    post:
      consumes:
      - application/json
      description: generate JWT token
      parameters:
      - description: Write Email and Password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JwtResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Generate JWT Token
      tags:
      - auth
  /singup/:
    post:
      consumes:
      - application/json
      description: create user
      parameters:
      - description: Write Email and Password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create User
      tags:
      - users
swagger: "2.0"